requirements.txt
plexapi~=4.14.0
pyyaml~=6.0.1